feat: Add default SQLite query in stats_query.txt
Left: Hour (in GM time), Right: fraction of active followers.

This SQLite query will return the % of your Tumblr followers that are
active at a particular hour.  The left # (from 0 to 23) is the hour in
Greenwich Mean time.  The right # (after the |) is the fraction of your
followers that have an idle time of 3600 seconds (1 hour) or less.